+++ codegen/Type/generateTypeStyles.ts
@@ -3,10 +3,23 @@ import {
   typographyConfig,
   TypographyConfig,
   cssFontFamilies,
+  fontWeights,
 } from 'eng/shared/design-system/codegen/configs/typographyConfig';
 
 import { DEFAULT_SCALE, scaleConfig } from '../configs/scaleConfig';
 import { calculateLetterSpacing, calculateMinFontSize, round } from './utils';
+
+type FontFamily = 'Graphik' | 'Inter';
+
+type TypographyStyles = {
+  fontFamily?: `${FontFamily}-${keyof typeof fontWeights}`;
+  // Mobile doesn't need font weight because the font family should include the weight
+  fontWeight?: typeof fontWeights[keyof typeof fontWeights];
+  fontSize: string | number;
+  lineHeight: string | number;
+  textTransform?: string;
+  letterSpacing?: string | number;
+};
 
 const calculateVariantStyle = (
   {
@@ -19,27 +32,27 @@ const calculateVariantStyle = (
     xHeight,
   }: TypographyConfig,
   scaleConversion: number,
-  { unitless }: { unitless: boolean }
+  { mobile }: { mobile: boolean } = { mobile: false }
 ) => {
   // Proposed font size by scale
   const rawFontSize = baseFontSize + scaleConversion;
   const roundedLineHeight = round(rawFontSize + leading);
-  const defaultFontSize = unitless ? rawFontSize : `${rawFontSize}px`;
-  const defaultLineHeight = unitless ? roundedLineHeight : `${roundedLineHeight}px`;
-  const defaultLetterSpacing = calculateLetterSpacing(rawFontSize, unitless);
+  const defaultFontSize = mobile ? rawFontSize : `${rawFontSize}px`;
+  const defaultLineHeight = mobile ? roundedLineHeight : `${roundedLineHeight}px`;
+  const defaultLetterSpacing = calculateLetterSpacing(rawFontSize, mobile);
 
   // check to see what minimum font sizes are for lowercase and upper case letters
   const minLowerCaseFontSize = Math.ceil(calculateMinFontSize(xHeight));
   const minLowercaseLineHeight = round(minLowerCaseFontSize + leading);
-  const minLowercaseLetterSpacing = calculateLetterSpacing(minLowerCaseFontSize, unitless);
+  const minLowercaseLetterSpacing = calculateLetterSpacing(minLowerCaseFontSize, mobile);
 
   const minUpperCaseFontSize = Math.ceil(calculateMinFontSize(1));
   const minUpperCaseLineHeight = round(minUpperCaseFontSize + leading);
-  const minUpperCaseLetterSpacing = calculateLetterSpacing(minUpperCaseFontSize, unitless);
+  const minUpperCaseLetterSpacing = calculateLetterSpacing(minUpperCaseFontSize, mobile);
 
   const sharedProps = {
-    fontWeight: unitless ? String(fontWeight) : fontWeight,
-    ...(unitless ? { fontFamily } : {}),
+    fontWeight: fontWeights[fontWeight],
+    fontFamily: `${fontFamily}-${fontWeight}` as const,
   };
 
   const lowercase = {
@@ -55,15 +68,15 @@ const calculateVariantStyle = (
   };
 
   const adjustedLowercase = {
-    fontSize: unitless ? minLowerCaseFontSize : `${minLowerCaseFontSize}px`,
-    lineHeight: unitless ? minLowercaseLineHeight : `${minLowercaseLineHeight}px`,
+    fontSize: mobile ? minLowerCaseFontSize : `${minLowerCaseFontSize}px`,
+    lineHeight: mobile ? minLowercaseLineHeight : `${minLowercaseLineHeight}px`,
     letterSpacing: minLowercaseLetterSpacing,
     ...sharedProps,
   };
 
   const adjustedUppercase = {
-    fontSize: unitless ? minUpperCaseFontSize : `${minUpperCaseFontSize}px`,
-    lineHeight: unitless ? minUpperCaseLineHeight : `${minUpperCaseLineHeight}px`,
+    fontSize: mobile ? minUpperCaseFontSize : `${minUpperCaseFontSize}px`,
+    lineHeight: mobile ? minUpperCaseLineHeight : `${minUpperCaseLineHeight}px`,
     textTransform: 'uppercase',
     letterSpacing: minUpperCaseLetterSpacing,
     ...sharedProps,
@@ -72,38 +85,45 @@ const calculateVariantStyle = (
   const lowercasePasses = rawFontSize >= minLowerCaseFontSize;
   const uppercasePasses = rawFontSize >= minUpperCaseFontSize;
 
+  let styles: TypographyStyles = adjustedUppercase;
   if (lowercasePasses) {
-    return lowercase;
+    styles = lowercase;
   } else if (!lowercasePasses && uppercasePasses && allowAllCaps) {
-    return allCaps;
+    styles = allCaps;
   } else if (!lowercasePasses && !allowAllCaps && !disableMinimums) {
-    return adjustedLowercase;
+    styles = adjustedLowercase;
   } else if (disableMinimums) {
-    return lowercase;
+    styles = lowercase;
   }
-  return adjustedUppercase;
+
+  if (mobile) {
+    // mobile will not be using letter spacing and font weight is baked into the font family name.
+    delete styles.letterSpacing;
+    delete styles.fontWeight;
+  } else {
+    // web doesn't need font-family in scale styles.
+    delete styles.fontFamily;
+  }
+
+  return styles;
 };
 
 // Codegen data
-export const typographyScaleMapWithoutUnits = mapValues(scaleConfig, scaleNumber =>
+export const typographyScaleMapForMobile = mapValues(scaleConfig, scaleNumber =>
   mapValues(typographyConfig, config =>
-    calculateVariantStyle(config, scaleNumber, { unitless: true })
+    calculateVariantStyle(config, scaleNumber, { mobile: true })
   )
 );
 
-export const typographyScaleMapWithUnits = mapValues(scaleConfig, scaleNumber =>
+export const typographyScaleMapForWeb = mapValues(scaleConfig, scaleNumber =>
   mapValues(typographyConfig, config =>
-    mapKeys(calculateVariantStyle(config, scaleNumber, { unitless: false }), (_, cssProperty) =>
-      kebabCase(cssProperty)
-    )
+    mapKeys(calculateVariantStyle(config, scaleNumber), (_, cssProperty) => kebabCase(cssProperty))
   )
 );
 
 export const typographyScaleMapWithCssVariables = mapValues(scaleConfig, (_, scaleName) => {
-  const textVariants = mapValues(
-    typographyScaleMapWithUnits[scaleName],
-    (stylesObject, variantName) =>
-      mapKeys(stylesObject, (_, cssProperty) => toCssVar(`${variantName}-${cssProperty}` as const))
+  const textVariants = mapValues(typographyScaleMapForWeb[scaleName], (stylesObject, variantName) =>
+    mapKeys(stylesObject, (_, cssProperty) => toCssVar(`${variantName}-${cssProperty}` as const))
   );
 
   return Object.entries(textVariants).reduce((prev, [_, next]) => {
@@ -115,7 +135,7 @@ export const typographyScaleMapWithCssVariables = mapValues(scaleConfig, (_, sca
 });
 
 export const typographyCss = mapValues(
-  typographyScaleMapWithUnits[DEFAULT_SCALE],
+  typographyScaleMapForWeb[DEFAULT_SCALE],
   (stylesObject, variantName) => ({
     ...mapValues(stylesObject, (_, cssProperty) =>
       toCssVarFn(`${variantName}-${cssProperty}` as const)