+++ theme/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- fix: Remove letting spacing from mobile Text component and correct font family by @hannah-jin [#12157](https://github.cbhq.net/mono/repo/pull/12157)
 - feat: Add spacing variables by @hannah-jin [#11916](https://github.cbhq.net/mono/repo/pull/11916)
 
 ## 0.0.4 (January 25th, 2021 3:30pm PST) [#11711](https://github.cbhq.net/mono/repo/pull/11711)